+++ app/src/main/java/com/agileburo/anytype/ui/page/modals/TurnIntoFragment.kt
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.agileburo.anytype.R
+import com.agileburo.anytype.core_ui.features.page.TurnIntoActionReceiver
 import com.agileburo.anytype.core_ui.features.page.modal.AddBlockOrTurnIntoAdapter
 import com.agileburo.anytype.core_ui.layout.SpacingItemDecoration
 import com.agileburo.anytype.core_ui.model.UiBlock
@@ -13,14 +14,14 @@ import com.agileburo.anytype.core_utils.ext.dimen
 import com.agileburo.anytype.core_utils.ui.BaseBottomSheetFragment
 import kotlinx.android.synthetic.main.fragment_add_block.*
 
-class AddBlockFragment : BaseBottomSheetFragment() {
+class TurnIntoFragment : BaseBottomSheetFragment() {
 
     companion object {
-        fun newInstance(): AddBlockFragment = AddBlockFragment()
+        fun newInstance(): TurnIntoFragment = TurnIntoFragment()
     }
 
     private val addBlockOrTurnIntoAdapter = AddBlockOrTurnIntoAdapter(
-        views = AddBlockOrTurnIntoAdapter.addBlockAdapterData(),
+        views = AddBlockOrTurnIntoAdapter.turnIntoAdapterData(),
         onUiBlockClicked = { type -> dispatchAndExit(type) }
     )
 
@@ -28,7 +29,7 @@ class AddBlockFragment : BaseBottomSheetFragment() {
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? = inflater.inflate(R.layout.fragment_add_block, container, false)
+    ): View? = inflater.inflate(R.layout.fragment_turn_into, container, false)
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -46,14 +47,10 @@ class AddBlockFragment : BaseBottomSheetFragment() {
     }
 
     private fun dispatchAndExit(block: UiBlock) {
-        (parentFragment as? AddBlockActionReceiver)?.onAddBlockClicked(block)
+        (parentFragment as? TurnIntoActionReceiver)?.onTurnIntoBlockClicked(block)
         dismiss()
     }
 
     override fun injectDependencies() {}
     override fun releaseDependencies() {}
-
-    interface AddBlockActionReceiver {
-        fun onAddBlockClicked(block: UiBlock)
-    }
 }