+++ core-ui/src/main/java/com/agileburo/anytype/core_ui/extensions/TextInputWidgetExtension.kt
@@ -1,0 +1,9 @@
+package com.agileburo.anytype.core_ui.extensions
+
+import com.agileburo.anytype.core_ui.widgets.text.TextInputWidget
+
+fun TextInputWidget.preserveSelection(block: () -> Unit) = synchronized(this) {
+    val selection = selectionStart..selectionEnd
+    block()
+    setSelection(selection.first, selection.last)
+}