+++ core-ui/src/main/java/com/agileburo/anytype/core_ui/features/page/BlockViewHolder.kt
@@ -30,7 +30,6 @@ import com.agileburo.anytype.core_ui.features.page.BlockViewDiffUtil.Companion.N
 import com.agileburo.anytype.core_ui.features.page.BlockViewDiffUtil.Companion.TEXT_CHANGED
 import com.agileburo.anytype.core_ui.features.page.BlockViewDiffUtil.Companion.TOGGLE_EMPTY_STATE_CHANGED
 import com.agileburo.anytype.core_ui.features.page.BlockViewDiffUtil.Payload
-import com.agileburo.anytype.core_ui.menu.TextBlockContextMenu
 import com.agileburo.anytype.core_ui.tools.DefaultSpannableFactory
 import com.agileburo.anytype.core_ui.tools.DefaultTextWatcher
 import com.agileburo.anytype.core_ui.widgets.text.EditorLongClickListener
@@ -38,7 +37,6 @@ import com.agileburo.anytype.core_ui.widgets.text.TextInputWidget
 import com.agileburo.anytype.core_utils.const.MimeTypes
 import com.agileburo.anytype.core_utils.ext.addDot
 import com.agileburo.anytype.core_utils.ext.dimen
-import com.agileburo.anytype.core_utils.ext.hideKeyboard
 import com.agileburo.anytype.core_utils.ext.imm
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.DataSource
@@ -100,30 +98,7 @@ sealed class BlockViewHolder(view: View) : RecyclerView.ViewHolder(view) {
         override val content: TextInputWidget = itemView.textContent
 
         init {
-            with(content) {
-                setSpannableFactory(DefaultSpannableFactory())
-                customSelectionActionModeCallback = TextBlockContextMenu(
-                    onTextColorClicked = { mode ->
-                        val start = content.selectionStart
-                        val end = content.selectionEnd
-                        content.hideKeyboard()
-                        onMarkupActionClicked(Markup.Type.TEXT_COLOR)
-                        mode.finish()
-                        content.setSelection(start, end)
-                        false
-                    },
-                    onBackgroundColorClicked = { mode ->
-                        val start = content.selectionStart
-                        val end = content.selectionEnd
-                        content.hideKeyboard()
-                        onMarkupActionClicked(Markup.Type.BACKGROUND_COLOR)
-                        mode.finish()
-                        content.setSelection(start, end)
-                        false
-                    },
-                    onMenuItemClicked = { onMarkupActionClicked(it) }
-                )
-            }
+            setup(onMarkupActionClicked)
         }
 
         fun bind(
@@ -417,30 +392,7 @@ sealed class BlockViewHolder(view: View) : RecyclerView.ViewHolder(view) {
         override val root: View = itemView
 
         init {
-            with(content) {
-                setSpannableFactory(DefaultSpannableFactory())
-                customSelectionActionModeCallback = TextBlockContextMenu(
-                    onTextColorClicked = { mode ->
-                        val start = content.selectionStart
-                        val end = content.selectionEnd
-                        content.hideKeyboard()
-                        onMarkupActionClicked(Markup.Type.TEXT_COLOR)
-                        mode.finish()
-                        content.setSelection(start, end)
-                        false
-                    },
-                    onBackgroundColorClicked = { mode ->
-                        val start = content.selectionStart
-                        val end = content.selectionEnd
-                        content.hideKeyboard()
-                        onMarkupActionClicked(Markup.Type.BACKGROUND_COLOR)
-                        mode.finish()
-                        content.setSelection(start, end)
-                        false
-                    },
-                    onMenuItemClicked = { onMarkupActionClicked(it) }
-                )
-            }
+            setup(onMarkupActionClicked)
         }
 
         fun bind(
@@ -538,30 +490,7 @@ sealed class BlockViewHolder(view: View) : RecyclerView.ViewHolder(view) {
         override val root: View = itemView
 
         init {
-            with(content) {
-                setSpannableFactory(DefaultSpannableFactory())
-                customSelectionActionModeCallback = TextBlockContextMenu(
-                    onTextColorClicked = { mode ->
-                        val start = content.selectionStart
-                        val end = content.selectionEnd
-                        content.hideKeyboard()
-                        onMarkupActionClicked(Markup.Type.TEXT_COLOR)
-                        mode.finish()
-                        content.setSelection(start, end)
-                        false
-                    },
-                    onBackgroundColorClicked = { mode ->
-                        val start = content.selectionStart
-                        val end = content.selectionEnd
-                        content.hideKeyboard()
-                        onMarkupActionClicked(Markup.Type.BACKGROUND_COLOR)
-                        mode.finish()
-                        content.setSelection(start, end)
-                        false
-                    },
-                    onMenuItemClicked = { onMarkupActionClicked(it) }
-                )
-            }
+            setup(onMarkupActionClicked)
         }
 
         fun bind(
@@ -638,30 +567,7 @@ sealed class BlockViewHolder(view: View) : RecyclerView.ViewHolder(view) {
         override val root: View = itemView
 
         init {
-            with(content) {
-                setSpannableFactory(DefaultSpannableFactory())
-                customSelectionActionModeCallback = TextBlockContextMenu(
-                    onTextColorClicked = { mode ->
-                        val start = content.selectionStart
-                        val end = content.selectionEnd
-                        content.hideKeyboard()
-                        onMarkupActionClicked(Markup.Type.TEXT_COLOR)
-                        mode.finish()
-                        content.setSelection(start, end)
-                        false
-                    },
-                    onBackgroundColorClicked = { mode ->
-                        val start = content.selectionStart
-                        val end = content.selectionEnd
-                        content.hideKeyboard()
-                        onMarkupActionClicked(Markup.Type.BACKGROUND_COLOR)
-                        mode.finish()
-                        content.setSelection(start, end)
-                        false
-                    },
-                    onMenuItemClicked = { onMarkupActionClicked(it) }
-                )
-            }
+            setup(onMarkupActionClicked)
         }
 
         fun bind(
@@ -741,30 +647,7 @@ sealed class BlockViewHolder(view: View) : RecyclerView.ViewHolder(view) {
         override val root: View = itemView
 
         init {
-            with(content) {
-                setSpannableFactory(DefaultSpannableFactory())
-                customSelectionActionModeCallback = TextBlockContextMenu(
-                    onTextColorClicked = { mode ->
-                        val start = content.selectionStart
-                        val end = content.selectionEnd
-                        content.hideKeyboard()
-                        onMarkupActionClicked(Markup.Type.TEXT_COLOR)
-                        mode.finish()
-                        content.setSelection(start, end)
-                        false
-                    },
-                    onBackgroundColorClicked = { mode ->
-                        val start = content.selectionStart
-                        val end = content.selectionEnd
-                        content.hideKeyboard()
-                        onMarkupActionClicked(Markup.Type.BACKGROUND_COLOR)
-                        mode.finish()
-                        content.setSelection(start, end)
-                        false
-                    },
-                    onMenuItemClicked = { onMarkupActionClicked(it) }
-                )
-            }
+            setup(onMarkupActionClicked)
         }
 
         fun bind(