+++ core-ui/src/main/java/com/agileburo/anytype/core_ui/features/page/TurnIntoActionReceiver.kt
@@ -1,0 +1,7 @@
+package com.agileburo.anytype.core_ui.features.page
+
+import com.agileburo.anytype.core_ui.model.UiBlock
+
+interface TurnIntoActionReceiver {
+    fun onTurnIntoBlockClicked(block: UiBlock)
+}